refactory query logic
