+++ contracts/ics721/src/lib.rs
@@ -2,6 +2,7 @@ pub mod contract;
 mod error;
 pub mod ibc;
 pub mod msg;
+pub mod query;
 pub mod state;
 mod test_constants;
 mod test_helpers;
